Move prepare_pixels helper functions to canvas_traits
