+++ components/canvas_traits/Cargo.toml
@@ -14,14 +14,17 @@ path = "lib.rs"
 webgl_backtrace = []
 
 [dependencies]
+byteorder = "1"
 cssparser = "0.25"
 euclid = "0.19"
 ipc-channel = "0.11"
-gleam = "0.6"
+gleam = "0.6.7"
+half = "1"
 lazy_static = "1"
 malloc_size_of = { path = "../malloc_size_of" }
 malloc_size_of_derive = { path = "../malloc_size_of_derive" }
 offscreen_gl_context = {version = "0.21", features = ["serde"]}
+pixels = {path = "../pixels"}
 serde = "1.0"
 serde_bytes = "0.10"
 servo_config = {path = "../config"}
