+++ components/canvas_traits/lib.rs
@@ -14,5 +14,6 @@ extern crate malloc_size_of_derive;
 extern crate serde;
 
 pub mod canvas;
+#[macro_use]
 pub mod webgl;
 mod webgl_channel;
