+++ components/script/Cargo.toml
@@ -53,7 +53,6 @@ enum-iterator = "0.2.0"
 euclid = "0.19"
 fnv = "1.0"
 gleam = "0.6"
-half = "1.0"
 headers-core = "0.0.1"
 headers-ext = "0.0.3"
 html5ever = "0.22"
